move popup to component
